requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.2.0
langchain-google-genai==0.1.0
faiss-cpu==1.7.4
pypdf==3.17.4
python-magic==0.4.27
reportlab==4.1.0
qrcode[pil]==7.4.2
